Refactor MainWindow into separate modules
