+++ ConcertActions.h
@@ -1,0 +1,4 @@
+#ifndef CONCERTACTIONS_H
+#define CONCERTACTIONS_H
+#include "mainwindow.h"
+#endif
