+++ Dialogs.h
@@ -1,0 +1,4 @@
+#ifndef DIALOGS_H
+#define DIALOGS_H
+#include "mainwindow.h"
+#endif
