+++ Export.h
@@ -1,0 +1,4 @@
+#ifndef EXPORT_H
+#define EXPORT_H
+#include "mainwindow.h"
+#endif
