+++ StudentActions.h
@@ -1,0 +1,4 @@
+#ifndef STUDENTACTIONS_H
+#define STUDENTACTIONS_H
+#include "mainwindow.h"
+#endif
